Tech Debt 2021-April: reduce compilation warnings for MME compilation
