+++ lte/gateway/c/oai/tasks/nas5g/src/ies/M5GAuthenticationResponseParameter.cpp
@@ -56,9 +56,9 @@ int AuthenticationResponseParameterMsg::
     EncodeAuthenticationResponseParameterMsg(
         AuthenticationResponseParameterMsg* response_parameter, uint8_t iei,
         uint8_t* buffer, uint32_t len) {
-  uint16_t* lenPtr;
   uint32_t encoded = 0;
 #ifdef HANDLE_POST_MVC
+  uint16_t* lenPtr;
   // Checking IEI and pointer
   CHECK_PDU_POINTER_AND_LENGTH_ENCODER(
       buffer, AUTHENTICATION_RESPONSE_PARAMETER_MIN_LEN, len);
